Move common  audio code into base class.
