+++ adafruit_circuitplayground/circuit_playground_base.py
@@ -36,6 +36,13 @@ CircuitPython base class for Circuit Playground.
 * Author(s): Kattni Rembor, Scott Shawcroft
 """
 
+import math
+import array
+import time
+try:
+    import audiocore
+except ImportError:
+    audiocore = audioio
 import adafruit_lis3dh
 import adafruit_thermistor
 import analogio
@@ -66,6 +73,9 @@ class Photocell:
 
 class CircuitPlaygroundBase:     # pylint: disable=too-many-public-methods
     """Circuit Playground base class."""
+
+    _audio_out = None
+
     def __init__(self):
         self._a = digitalio.DigitalInOut(board.BUTTON_A)
         self._a.switch_to_input(pull=digitalio.Pull.DOWN)
@@ -101,6 +111,13 @@ class CircuitPlaygroundBase:     # pylint: disable=too-many-public-methods
         self._int1 = digitalio.DigitalInOut(board.ACCELEROMETER_INTERRUPT)
         self._lis3dh = adafruit_lis3dh.LIS3DH_I2C(self._i2c, address=0x19, int1=self._int1)
         self._lis3dh.range = adafruit_lis3dh.RANGE_8_G
+
+        # Define audio:
+        self._speaker_enable = digitalio.DigitalInOut(board.SPEAKER_ENABLE)
+        self._speaker_enable.switch_to_output(value=False)
+        self._sample = None
+        self._sine_wave = None
+        self._sine_wave_sample = None
 
         # Initialise tap:
         self._detect_taps = 1
@@ -811,3 +828,178 @@ class CircuitPlaygroundBase:     # pylint: disable=too-many-public-methods
     @red_led.setter
     def red_led(self, value):
         self._led.value = value
+
+    @staticmethod
+    def _sine_sample(length):
+        tone_volume = (2 ** 15) - 1
+        shift = 2 ** 15
+        for i in range(length):
+            yield int(tone_volume * math.sin(2*math.pi*(i / length)) + shift)
+
+    def _generate_sample(self, length=100):
+        if self._sample is not None:
+            return
+        self._sine_wave = array.array("H", self._sine_sample(length))
+        self._sample = self._audio_out(board.SPEAKER)
+        self._sine_wave_sample = audiocore.RawSample(self._sine_wave)
+
+    def play_tone(self, frequency, duration):
+        """ Produce a tone using the speaker. Try changing frequency to change
+        the pitch of the tone.
+
+        :param int frequency: The frequency of the tone in Hz
+        :param float duration: The duration of the tone in seconds
+
+        .. image :: ../docs/_static/speaker.jpg
+          :alt: Onboard speaker
+
+        To use with the Circuit Playground Express:
+
+        .. code-block:: python
+
+            from adafruit_circuitplayground.express import cpx
+
+            cpx.play_tone(440, 1)
+
+        To use with the Circuit Playground Bluefruit:
+
+        .. code-block:: python
+
+            from adafruit_circuitplayground.bluefruit import cpb
+
+            cpb.play_tone(440, 1)
+        """
+        # Play a tone of the specified frequency (hz).
+        self.start_tone(frequency)
+        time.sleep(duration)
+        self.stop_tone()
+
+    def start_tone(self, frequency):
+        """ Produce a tone using the speaker. Try changing frequency to change
+        the pitch of the tone.
+
+        :param int frequency: The frequency of the tone in Hz
+
+        .. image :: ../docs/_static/speaker.jpg
+          :alt: Onboard speaker
+
+        To use with the Circuit Playground Express:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.express import cpx
+
+             while True:
+                 if cpx.button_a:
+                     cpx.start_tone(262)
+                 elif cpx.button_b:
+                     cpx.start_tone(294)
+                 else:
+                     cpx.stop_tone()
+
+        To use with the Circuit Playground Bluefruit:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.bluefruit import cpb
+
+             while True:
+                 if cpb.button_a:
+                     cpb.start_tone(262)
+                 elif cpb.button_b:
+                     cpb.start_tone(294)
+                 else:
+                     cpb.stop_tone()
+        """
+        self._speaker_enable.value = True
+        length = 100
+        if length * frequency > 350000:
+            length = 350000 // frequency
+        self._generate_sample(length)
+        # Start playing a tone of the specified frequency (hz).
+        self._sine_wave_sample.sample_rate = int(len(self._sine_wave) * frequency)
+        if not self._sample.playing:
+            self._sample.play(self._sine_wave_sample, loop=True)
+
+    def stop_tone(self):
+        """ Use with start_tone to stop the tone produced.
+
+        .. image :: ../docs/_static/speaker.jpg
+          :alt: Onboard speaker
+
+        To use with the Circuit Playground Express:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.express import cpx
+
+             while True:
+                 if cpx.button_a:
+                     cpx.start_tone(262)
+                 elif cpx.button_b:
+                     cpx.start_tone(294)
+                 else:
+                     cpx.stop_tone()
+
+        To use with the Circuit Playground Bluefruit:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.bluefruit import cpb
+
+             while True:
+                 if cpb.button_a:
+                     cpb.start_tone(262)
+                 elif cpb.button_b:
+                     cpb.start_tone(294)
+                 else:
+                     cpb.stop_tone()
+        """
+        # Stop playing any tones.
+        if self._sample is not None and self._sample.playing:
+            self._sample.stop()
+            self._sample.deinit()
+            self._sample = None
+        self._speaker_enable.value = False
+
+    def play_file(self, file_name):
+        """ Play a .wav file using the onboard speaker.
+
+        :param file_name: The name of your .wav file in quotation marks including .wav
+
+        .. image :: ../docs/_static/speaker.jpg
+          :alt: Onboard speaker
+
+        To use with the Circuit Playground Express:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.express import cpx
+
+             while True:
+                 if cpx.button_a:
+                     cpx.play_file("laugh.wav")
+                 elif cpx.button_b:
+                     cpx.play_file("rimshot.wav")
+
+        To use with the Circuit Playground Bluefruit:
+
+        .. code-block:: python
+
+             from adafruit_circuitplayground.bluefruit import cpb
+
+             while True:
+                 if cpb.button_a:
+                     cpb.play_file("laugh.wav")
+                 elif cpb.button_b:
+                     cpb.play_file("rimshot.wav")
+        """
+        # Play a specified file.
+        self.stop_tone()
+        self._speaker_enable.value = True
+        with self._audio_out(board.SPEAKER) as audio:
+            wavefile = audiocore.WaveFile(open(file_name, "rb"))
+            audio.play(wavefile)
+            while audio.playing:
+                pass
+        self._speaker_enable.value = False
